Move some commont functionality into bench_results and utils
