+++ tools/performance/engine-benchmarks/bench_tool/utils.py
@@ -1,0 +1,31 @@
+import shutil
+import tempfile
+from typing import List, Set
+
+from bench_tool import JobReport
+
+
+class WithTempDir:
+    def __init__(self, prefix: str):
+        self.prefix = prefix
+        self.temp_dir = None
+
+    def __enter__(self):
+        self.temp_dir = tempfile.mkdtemp(prefix=self.prefix)
+        return self.temp_dir
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        shutil.rmtree(self.temp_dir, ignore_errors=True)
+
+
+def gather_all_bench_labels(job_reports: List[JobReport]) -> Set[str]:
+    """
+    Iterates through all the job reports and gathers all the benchmark labels
+    found. Note that every job report can have a different set of benchmark labels.
+    :return: List of benchmark labels.
+    """
+    all_labels = set()
+    for job_report in job_reports:
+        for labels in job_report.label_score_dict.keys():
+            all_labels.add(labels)
+    return all_labels
